+++ dlokiosk/src/main/java/com/dlohaiti/dlokiosk/domain/Register.java
@@ -26,7 +26,7 @@ public class Register {
 
     public Receipt checkout(ShoppingCart cart) {
         List<LineItem> lineItems = buildLineItemsFrom(cart);
-        int totalGallons = 0;
+        Double totalGallons = 0.0;
         for (Product product : cart.getProducts()) {
             totalGallons += product.getGallons() * product.getQuantity();
         }
